started working on trailers

+++ thetrailers/lib/pages/movies/movies_index.dart
@@ -1,8 +1,9 @@
+import 'package:flutter/material.dart';
 import 'package:thetrailers/models/movie.dart';
 import 'package:thetrailers/pages/movies/movies_create.dart';
 import 'package:thetrailers/pages/movies/movies_edit.dart';
+import 'package:thetrailers/pages/movies/movies_trailers.dart';
 import 'package:thetrailers/services/movie_service.dart';
-import 'package:flutter/material.dart';
 
 class MoviesIndex extends StatefulWidget {
   const MoviesIndex({Key? key}) : super(key: key);
@@ -12,6 +13,13 @@ class MoviesIndex extends StatefulWidget {
 }
 
 class _MoviesIndexState extends State<MoviesIndex> {
+  Future<void> _createMovie(BuildContext context) async {
+    await Navigator.of(context).push(
+      MaterialPageRoute(builder: (context) => const MoviesCreate()),
+    );
+    setState(() {});
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -30,7 +38,7 @@ class _MoviesIndexState extends State<MoviesIndex> {
           return _moviesIndex(snapshot.data!, context);
         },
       ),
-floatingActionButton: FloatingActionButton(
+      floatingActionButton: FloatingActionButton(
         child: const Icon(Icons.add),
         onPressed: () => _createMovie(context),
       ),
@@ -41,32 +49,33 @@ floatingActionButton: FloatingActionButton(
     return ListView.builder(
       itemCount: list.length,
       itemBuilder: (context, index) {
-        return ListTile(
-          title: Row(
-            children: [
-              Expanded(child: Text(list[index].title)),
-              _editMovie(list[index], context),
-              _deleteMovie(list[index].id, context)
-            ],
+        return Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: ListTile(
+            leading: _movieTrailers(list[index], context),
+            title: Text(list[index].title),
+            trailing: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                _editMovie(list[index], context),
+                const SizedBox(width: 10),
+                _deleteMovie(list[index].id, context),
+              ],
+            ),
+            subtitle:
+                Text('Aantal trailers: ${list[index].trailers?.length ?? 0}'),
           ),
         );
       },
     );
   }
 
-  Future<void> _createMovie(BuildContext context) async {
-    await Navigator.of(context).push(MaterialPageRoute(
-      builder: (context) => const MoviesCreate(),
-    ));
-    setState(() {});
-  }
-
   Widget _editMovie(Movie movie, BuildContext context) {
     return GestureDetector(
       onTap: () async {
-        await Navigator.of(context).push(MaterialPageRoute(
-          builder: (context) => MoviesEdit(movie: movie),
-        ));
+        await Navigator.of(context).push(
+          MaterialPageRoute(builder: (context) => MoviesEdit(movie: movie)),
+        );
         setState(() {});
       },
       child: const Icon(Icons.edit),
@@ -75,30 +84,64 @@ floatingActionButton: FloatingActionButton(
 
   Widget _deleteMovie(int id, BuildContext context) {
     return GestureDetector(
-      onTap: () async {
-        bool gelukt = await MovieService().delete(id);
-        if (gelukt) {
-          setState(() {});
-        } else {
-          if (context.mounted) {
-            showDialog(
-              context: context,
-              builder: (context) {
-                return AlertDialog(
-                  title: const Text('Movies - Delete'),
-                  content: const Text('Verwijderen is mislukt'),
-                  actions: [
-                    TextButton(
-                        onPressed: () => Navigator.of(context).pop(),
-                        child: const Text('Ok'))
-                  ],
-                );
-              },
+      onTap: () {
+        showDialog(
+          context: context,
+          builder: (context) {
+            return AlertDialog(
+              title: const Text('Movies - Delete'),
+              content:
+                  const Text('Are you sure you want to delete this movie?'),
+              actions: [
+                TextButton(
+                  onPressed: () async {
+                    bool success = await MovieService().delete(id);
+                    if (success) {
+                      Navigator.of(context).pop(); // Close the dialog
+                      setState(() {}); // Update the UI
+                    } else {
+                      Navigator.of(context).pop(); // Close the dialog
+                      showDialog(
+                        context: context,
+                        builder: (context) {
+                          return AlertDialog(
+                            title: const Text('Movies - Delete'),
+                            content: const Text('Deletion failed.'),
+                            actions: [
+                              TextButton(
+                                onPressed: () => Navigator.of(context).pop(),
+                                child: const Text('OK'),
+                              ),
+                            ],
+                          );
+                        },
+                      );
+                    }
+                  },
+                  child: const Text('Delete'),
+                ),
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(),
+                  child: const Text('Cancel'),
+                ),
+              ],
             );
-          }
-        }
+          },
+        );
       },
       child: const Icon(Icons.delete_outline),
+    );
+  }
+
+    Widget _movieTrailers(Movie movie, BuildContext context) {
+    return GestureDetector(
+      onTap: () async {
+        await Navigator.of(context).push(MaterialPageRoute(
+          builder: (context) => MoviesTrailersPage(movie: movie),
+        ));
+        setState(() {});
+      },
+      child: const Icon(Icons.account_box_outlined),
     );
   }
 }
